requirements.txt
kubernetes>=28.1.0
anthropic>=0.39.0
jira>=3.8.0
pydantic>=2.5.0
pyyaml>=6.0.1
tenacity>=8.2.0
structlog>=24.1.0
